DAPI-1354: move score evolution query from application to infrastructure
