+++ src/Akeneo/Pim/Automation/DataQualityInsights/back/Infrastructure/ProductGrid/AddProductScoreProperty.php
@@ -16,7 +16,7 @@ use Akeneo\Pim\Enrichment\Component\Product\Grid\ReadModel\AdditionalProperty;
  * @copyright 2020 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-final class AddProductScoreProperty  implements AddAdditionalProductProperties
+final class AddProductScoreProperty implements AddAdditionalProductProperties
 {
     private GetLatestProductScoresQueryInterface $getProductScores;
