+++ src/Akeneo/Pim/Automation/DataQualityInsights/back/Infrastructure/ProductGrid/QualityScoreFilter.php
@@ -11,7 +11,7 @@ use Oro\Bundle\FilterBundle\Filter\ChoiceFilter;
  * @copyright 2020 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-final class QualityScoreFilter  extends ChoiceFilter
+final class QualityScoreFilter extends ChoiceFilter
 {
     /**
      * {@inheritdoc}
